refactor(messages): extract idle acknowledgement text detection
Code motion only, no behaviour change.

looksLikeIdleAckOnlyText and its normaliser and vocabulary lived inside
messages.js, where only the placeholder-task-ref path used them. The next
commit needs the same predicate inside messageStore.js, and a message store
that requires messages.js would invert the dependency between the two.

They move verbatim into internal/idleAckText.js. messageStore.js requires the
new module and re-exports the predicate, and messages.js takes it from there,
so the ack vocabulary keeps exactly one import surface for callers. The two
declarations are reordered so the normaliser precedes the phrase set, which the
next commit needs in order to normalise its lists at module load.

messages.js drops from 247 to 172 lines; the new module is 85.
